Add files via upload
I went and added code to blink LEDs at specific times for different reasons. Changed a function name to allow me to do more. LEDs blink green 3 times if the fingerprint scanner is detected; red otherwise. The scanner will blink for 2 seconds if the finger scanned is in the database; red otherwise.