(20 May 2019) Fixed issues with array storage and database restart
Database restart: The issue with the database restart has been fixed. Previously, the system required an owner [ID 0] to be present and authenticate themselves before any fingerprint could be entered. However, this entered a state of impossibility when the database was deleted and no owner was entered. The program now allows for anybody to enter when no one is in the system and it will automatically assign them to ID 0 - the owner. Upon entry, the program will now require the owner to verify themselves before adding anyone else to the system.

Array storage: The array that is being used to monitor what IDs were currently being stored was rerewriting everything within the system into the array, with every successful new entry. For example, if the array already had ID 0, 1, 2 in it and I added ID 35, the array would now hold 0, 1, 2, 0, 1, 2, 35. If I added 76 next, the array would hold the values: 0, 1, 2, 0, 1, 2, 35, 76. This would cause (after a small # of entries) the filling of the entire array that is needed to verify uniqueness of IDs. This problem has been resolved and will now show up as (assuming the last value example given above) as: 0, 1, 2, 35, 76. If I add 4 it will now show up as: 0, 1, 2, 3, 4, 35, 76.