(22 May 2019) Test files for Fingerprint sensor.
Basic code to make the fingerprint sensor operate. Has not been tested (still needs to), but has been checked multiple times, with multiple people, to ensure that it works. Needs to run within the Raspberry Pi Python file. Will not run in Spyder due to integrated I/O being involved.

This code will not be the code that is used in the final product, it is just code to ensure that different states can use the data and that data is read properly. The code has been created from a lot of research and from code that was in our repository and potentially altered at the time of pulling the .zip file.

The code is in .txt files, but can easily be transferred to the Python file.