(23 May 2019)
This file has not been fully tested yet, but has passed the recent tests, thus far. It definitely needs to be run via a program vs. typing through the terminal.

A file with auto writes back to the Elegoo is the goal, but the as of now, the fingerprint scanner turns on/off upon request. The buttons have been eliminated. The features from before still stand.

An internal menu will have the following options:
1) Read/Verify
2) Add new user
3) Delete Individual
4) Wrong Order
5) Delete database

All of these will be controlled in Python with write values on the Raspberry Pi --> Elegoo.