(27 May 2019) Fixed looping issue with adding fingerprint - Empty database
The issue with the infinitely looping adding fingerprint state when adding a new user has been fixed. The issue that arose was the fingerprint scanner would allow for a fingerprint to be added when the database was empty and assign it the value of zero. The database, however, would never break out of that state and loop back in allowing for an infinite amount of scans, all with the value of 0. 

This caused improper communication to occur via UART and the system could never move on properly. This issue has now been fixed. After the fingerprint is added a write value of '6' to the Elegoo from the Raspberry Pi, via UART, will now turn the sensor off and allow for picking of a new command.