(28 May 2019) Test file for fingerprint sensor - Python
This is a test code to send and receive commands via UART between the Raspberry Pi and Elegoo. This allows for testing of all features that the fingerprint sensor has, all through a menu in Python.

The test turns the fingerprint sensor on at appropriate times, sending commands to find the correct state on the Elegoo, then sends another command to turn off the sensor when the task is complete.