(29 May 2019) Core code with fingerprint integration
This file has the test code that I created for Raspberry Pi and Elegoo communication for the fingerprint scanner. That code has been pulled apart and integrated inside of the core file. The data is saved to a .txt file (as of now), communication is established via UART, and an image is taken after three incorrect attempts in the phase where one tries and access the barcode phase.

Once an image is taken, it will be run against a file that has images of faces, using OpenCV. If the system detects a face in the newly taken image, a box will be placed around that detection.