(3 June 2019) Fixed two issues
There were two issues with the code on the Elegoo for the fingerprint sensor.

Issue 1 - Fixed: It would write via UART an old value, using an old method of confirmation. This old value was the negative version of the unique ID that was individually deleted from the fingerprint sensor's internal database.

Issue 2 - Fixed: After adding the first fingerprint to the database, shutting the sensor off would cause the value of "Features could not be found" would be sent via UART. This would cause error and lock up of the system on the Raspberry Pi. This no longer sends that via UART.

Benefit of fixes: The code should be able to add a first fingerprint and shut the sensor off without locking the system (requiring a restart). The system now will write a value and it will not lower the value of the counter to 2, instead of 3. The counter is responsible for activating the camera sub-system.