Camera subsystem integration proof
This program is proof that the camera subsystem is integrated with our core code and does work with python3. It is on a different RPi as the original RPi was giving many issues with installing openCV. Because of this, the database is mocked with a database.txt file and there is no website to update on this new, test Pi.

The purpose of using a new RPi w/o the database and website is to prove that the camera subsystem is integrated and works within our code. This RPi allows the picamera to be used and has been set up with openCV and works within testing.